Move template registry to echo handlers
templates are managed there
