+++ echo_handlers.go
@@ -2,8 +2,10 @@ package main
 
 import (
 	"embed"
+	"errors"
 	"fmt"
 	"html/template"
+	"io"
 	"io/fs"
 	"net/http"
 	"strings"
@@ -23,6 +25,20 @@ var embeddedAssets embed.FS
 
 //go:embed views/*
 var embeddedViews embed.FS
+
+type TemplateRegistry struct {
+	templates map[string]*template.Template
+}
+
+// Implement e.Renderer interface
+func (t *TemplateRegistry) Render(w io.Writer, name string, data interface{}, c echo.Context) error {
+	tmpl, ok := t.templates[name]
+	if !ok {
+		err := errors.New("Template not found -> " + name)
+		return err
+	}
+	return tmpl.ExecuteTemplate(w, "layout.html", data)
+}
 
 func (svc *Service) RegisterSharedRoutes(e *echo.Echo) {
